mysite/requirements.txt
aldryn-apphooks-config==0.5.3
aldryn-boilerplates==0.8.0
aldryn-categories==1.2.0
aldryn-common==1.0.5
aldryn-people==2.2.0
aldryn-translation-tools==0.3.0
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
autopep8==1.5
Babel==2.8.0
backcall==0.1.0
backport-collections==0.1
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cmsplugin-filer==1.1.3
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.12
django-allauth==0.41.0
django-appconf==1.0.4
django-appdata==0.2.2
django-classy-tags==1.0.0
django-cms==3.7.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filer==1.7.0
django-formtools==2.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-markdown==0.8.4
django-mptt==0.11.0
django-parler==2.0.1
django-phonenumber-field==1.3.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-sortedm2m==3.0.0
django-taggit==1.2.0
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.5.0
djangocms-column==1.9.0
djangocms-file==2.4.0
djangocms-forms==0.2.5
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.2.1
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
easy-thumbnails==2.7
entrypoints==0.3
flake8==3.7.9
hashids==1.2.0
html5lib==1.0.1
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jsonfield==3.1.0
lxml==4.5.0
Markdown==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
parso==0.6.2
pathspec==0.7.0
phonenumbers==8.12.1
phonenumberslite==8.12.1
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.5
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
text-unidecode==1.3
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
YURL==1.0.0
